+++ core/src/test/java/org/nvip/plugfest/tooling/translator/utils/Utils.java
@@ -1,6 +1,9 @@
 package org.nvip.plugfest.tooling.translator.utils;
 
+import org.nvip.plugfest.tooling.sbom.Component;
 import org.nvip.plugfest.tooling.sbom.SBOM;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
 /**
  * File: Utils.java
@@ -13,6 +16,7 @@ public class Utils {
     /**
      * Helper method to check that SBOM metadata does not contain app tools
      * @param sbom to check
+     * @return whether metadata contains AppTools or not
      */
     public static boolean checkForAppTools(SBOM sbom) {
         for (String m: sbom.getMetadata().values()
@@ -21,6 +25,19 @@ public class Utils {
                 return true;
         }
         return false;
+    }
+
+    /**
+     * Helper method to check that SBOMs that should
+     * have no licenses do not have that field with any licenses added
+     * @param sbom to check
+     * @return true if SBOM contains no licenses throughout its components
+     */
+    public static boolean noLicensesCheck(SBOM sbom){
+        for (Component c: sbom.getAllComponents())
+            if(c.getLicenses().size() > 0)
+                return false;
+        return true;
     }
 
 }