+++ changelog.md
@@ -19,12 +19,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   - `HasExtractedLicensesTest` - check if there are any extracted licenses not on the SPDX license list
   - `ExtractedLicenseMinElementTest` - check any extracted licenses for the required fields: LicenseName, LicenseID, 
      LicenseCrossReference 
+- Method in Utils.java to ensure components with no licenses have no null components
 - Added additional exceptions thrown in `TranslatorCDXXML` to increase verbosity of error messages.
 
 ### Changed
 
 ### Fixed
 - Fixed broken SBOM file transfers by adding back the `@RequestBody` decorators to all `SBOMFile` API request parameters.
+- Fixed issue with CDXJSON translator regarding translating null license objects to components without licenses
 - Fixed errors caused by `TranslatorPlugFest.getTranslator()` method incorrectly determining filetype.
   - We now check for file extension only, since SPDX JSON and XML are not supported translators.